perf: improve calc tree
