+++ src/redux/currentApp/executionTree/dependencies/dependenciesListener.ts
@@ -1,9 +1,12 @@
 import { isAnyOf, Unsubscribe } from "@reduxjs/toolkit"
 import { componentsActions } from "@/redux/currentApp/editor/components/componentsSlice"
 import { getAllComponentDisplayNameMapProps } from "@/redux/currentApp/editor/components/componentsSelector"
-import { generateDependencies } from "@/utils/generators/generateDependenciesMap"
 import { dependenciesActions } from "@/redux/currentApp/executionTree/dependencies/dependenciesSlice"
 import { AppListenerEffectAPI, AppStartListening } from "@/store"
+import dependenciesTreeWorker from "@/utils/worker/dependenciesTreeWorker?worker"
+
+// TODO: need a channel
+export const worker = new dependenciesTreeWorker()
 
 async function handleUpdateDependencies(
   action: unknown,
@@ -12,10 +15,15 @@ async function handleUpdateDependencies(
   const rootState = listenerApi.getState()
   const displayNameMapProps = getAllComponentDisplayNameMapProps(rootState)
   if (!displayNameMapProps) return
-  const inverseDependencies = generateDependencies(displayNameMapProps)
-  listenerApi.dispatch(
-    dependenciesActions.setDependenciesReducer(inverseDependencies),
-  )
+  worker.postMessage({
+    action: "GENERATE_DEPENDENCIES",
+    displayNameMapProps: displayNameMapProps,
+  })
+  worker.onmessage = (e) => {
+    const { data } = e
+    const { result = {} } = data
+    listenerApi.dispatch(dependenciesActions.setDependenciesReducer(result))
+  }
 }
 
 export function setupDependenciesListeners(
