+++ src/utils/evaluateDynamicString/dynamicConverter.ts
@@ -73,7 +73,12 @@ export const getDynamicValue = (
   if (stringSnippets.length) {
     const values = jsSnippets.map((jsSnippet, index) => {
       if (jsSnippet) {
-        return evalScript(jsSnippet, dataTree, isTriggerBased)
+        console.log("jsSnippet", jsSnippet)
+        console.log("dataTree", dataTree)
+        console.log("isTriggerBased", isTriggerBased)
+        const result = evalScript(jsSnippet, dataTree, isTriggerBased)
+        console.log("result", result)
+        return result
       } else {
         return stringSnippets[index]
       }
